Refactor general MarkRenderer with MarkShader trait
