+++ vega-wgpu-renderer/src/renderers/mod.rs
@@ -1,11 +1,5 @@
-use crate::renderers::rect::RectMarkRenderer;
-use crate::renderers::symbol::SymbolMarkRenderer;
 pub mod canvas;
+mod mark;
 pub mod rect;
 pub mod symbol;
 pub mod vertex;
-
-pub enum MarkRenderer {
-    Symbol(SymbolMarkRenderer),
-    Rect(RectMarkRenderer),
-}
